new ChanSwitch view fix playback failed because of substring display time when pausing

+++ neteasecloudmusic/src/main/java/moe/lz233/meizugravity/cloudmusic/logic/dao/BaseDao.kt
@@ -3,7 +3,7 @@ package moe.lz233.meizugravity.cloudmusic.logic.dao
 import moe.lz233.meizugravity.cloudmusic.App
 
 object BaseDao {
-    var soundQuality: Long
-        get() = App.sp.getLong("soundQuality", 999000)
-        set(value) = App.editor.putLong("soundQuality", value).apply()
+    var soundQuality: Int
+        get() = App.sp.getInt("soundQuality", 999000)
+        set(value) = App.editor.putInt("soundQuality", value).apply()
 }
